Fixed game simulation still happening while in shop screen, removed move indicator

+++ video.c
@@ -161,10 +161,10 @@ void InitVideo()
 }
 void DrawRoundedRect(Rect r, ALLEGRO_COLOR color, bool filled)
 {
-    float x = (r.x); 
-    float y = (r.y); 
-    float x2 = (r.x + r.w);
-    float y2 = (r.y + r.h);
+    int x = (r.x); 
+    int y = (r.y); 
+    int x2 = (r.x + r.w);
+    int y2 = (r.y + r.h);
     //getting a precise output on mac just isnt working 
     //so here's a TODO: fix this... eventually
     #ifdef __APPLE__
